chore(broker): split route files into schemas/handlers/index per domain
Each route module (events, memory, sessions, stream, traces) is now a
three-file package: schemas.ts for zod schemas and derived types,
handlers.ts for the handle* business-logic functions, and index.ts for
the router factory that wires Express routes to those handlers.

The circular dep between trace-broker and traces route introduced in the
previous commit is resolved here: spanMatchesSessionId was already moved
into the broker; handlers.ts imports it from there.

StreamError moves to stream/schemas.ts so handlers.ts can reference it
without re-declaring it. CRD-shadow types (SessionEventData,
GetSessionsQuery, etc.) stay in their respective broker or schema files.
