+++ services/ark-broker/ark-broker/src/http/routes/memory/handlers.ts
@@ -1,0 +1,108 @@
+import type {Request, Response} from 'express';
+import {MemoryBroker} from '@ark-broker/brokers/memory-broker.js';
+import {streamSSE} from '../../sse.js';
+import {
+  parsePaginationParams,
+  PaginationError,
+  PaginatedList,
+} from '@ark-broker/brokers/pagination.js';
+import {sendPaginationError, sendInternalError} from '../errors.js';
+
+interface MessageItem {
+  timestamp: string;
+  conversation_id: string;
+  query_id: string;
+  message: unknown;
+  sequence: number;
+}
+
+export function handleStreamingMessages(
+  req: Request,
+  res: Response,
+  memory: MemoryBroker,
+  conversationId: string | undefined,
+  cursor: number | undefined
+): void {
+  req.log.info({cursor}, 'starting SSE stream for all messages');
+
+  let replayItems: MessageItem[] | undefined;
+  if (cursor !== undefined) {
+    let items = memory.all().filter((item) => item.sequenceNumber > cursor);
+    if (conversationId) {
+      items = items.filter(
+        (item) => item.data.conversationId === conversationId
+      );
+    }
+    replayItems = items.map((item) => ({
+      timestamp: item.timestamp.toISOString(),
+      conversation_id: item.data.conversationId,
+      query_id: item.data.queryId,
+      message: item.data.message,
+      sequence: item.sequenceNumber,
+    }));
+  }
+
+  streamSSE({
+    res,
+    req,
+    logger: req.log,
+    tag: 'MESSAGES',
+    itemName: 'messages',
+    subscribe: (callback) =>
+      memory.subscribe((item) => {
+        callback({
+          timestamp: item.timestamp.toISOString(),
+          conversation_id: item.data.conversationId,
+          query_id: item.data.queryId,
+          message: item.data.message,
+          sequence: item.sequenceNumber,
+        });
+      }),
+    filter: conversationId
+      ? (msg: unknown): boolean =>
+          (msg as {conversation_id: string}).conversation_id === conversationId
+      : undefined,
+    replayItems,
+  });
+}
+
+export function handlePaginatedMessages(
+  req: Request,
+  res: Response,
+  memory: MemoryBroker,
+  conversationId: string | undefined,
+  queryId: string | undefined
+): void {
+  try {
+    const params = parsePaginationParams(req.query as Record<string, unknown>);
+
+    const filters = {
+      conversationId: conversationId || undefined,
+      queryId: queryId || undefined,
+    };
+
+    const result = memory.paginate(params, filters);
+
+    const response: PaginatedList<MessageItem> = {
+      items: result.items.map((item) => ({
+        timestamp: item.timestamp.toISOString(),
+        conversation_id: item.data.conversationId,
+        query_id: item.data.queryId,
+        message: item.data.message,
+        sequence: item.sequenceNumber,
+      })),
+      total: result.total,
+      hasMore: result.hasMore,
+      nextCursor: result.nextCursor,
+    };
+
+    res.json(response);
+  } catch (error) {
+    if (error instanceof PaginationError) {
+      sendPaginationError(res, error, req.id);
+      return;
+    }
+    req.log.error({err: error}, 'failed to get messages');
+    sendInternalError(res, req.id);
+  }
+}
