+++ services/ark-broker/ark-broker/src/http/routes/memory/index.ts
@@ -1,143 +1,16 @@
 import {Router} from 'express';
-import type {Request, Response} from 'express';
-import {
-  sendValidationError,
-  sendPaginationError,
-  sendInternalError,
-} from './errors.js';
 import {randomUUID} from 'crypto';
-import {z} from 'zod';
 import {MemoryBroker} from '@ark-broker/brokers/memory-broker.js';
 import {SessionsBroker} from '@ark-broker/brokers/sessions-broker.js';
-import {streamSSE} from '../sse.js';
+import {sendValidationError, sendInternalError} from '../errors.js';
 import {
-  parsePaginationParams,
-  PaginationError,
-  PaginatedList,
-} from '@ark-broker/brokers/pagination.js';
-
-const postMessagesBodySchema = z.object({
-  conversation_id: z.string(),
-  query_id: z.string(),
-  messages: z.array(z.unknown()),
-});
-type PostMessagesBody = z.infer<typeof postMessagesBodySchema>;
-
-const getMessagesQuerySchema = z.object({
-  watch: z
-    .enum(['true', 'false'])
-    .transform((v) => v === 'true')
-    .optional(),
-  conversation_id: z.string().optional(),
-  query_id: z.string().optional(),
-  cursor: z.coerce.number().int().nonnegative().optional(),
-});
-type GetMessagesQuery = z.infer<typeof getMessagesQuerySchema>;
-type GetMessagesQueryRaw = {
-  watch?: 'true' | 'false';
-  conversation_id?: string;
-  query_id?: string;
-  cursor?: string;
-};
-
-interface MessageItem {
-  timestamp: string;
-  conversation_id: string;
-  query_id: string;
-  message: unknown;
-  sequence: number;
-}
-
-function handleStreamingMessages(
-  req: Request,
-  res: Response,
-  memory: MemoryBroker,
-  conversationId: string | undefined,
-  cursor: number | undefined
-): void {
-  req.log.info({cursor}, 'starting SSE stream for all messages');
-
-  let replayItems: MessageItem[] | undefined;
-  if (cursor !== undefined) {
-    let items = memory.all().filter((item) => item.sequenceNumber > cursor);
-    if (conversationId) {
-      items = items.filter(
-        (item) => item.data.conversationId === conversationId
-      );
-    }
-    replayItems = items.map((item) => ({
-      timestamp: item.timestamp.toISOString(),
-      conversation_id: item.data.conversationId,
-      query_id: item.data.queryId,
-      message: item.data.message,
-      sequence: item.sequenceNumber,
-    }));
-  }
-
-  streamSSE({
-    res,
-    req,
-    logger: req.log,
-    tag: 'MESSAGES',
-    itemName: 'messages',
-    subscribe: (callback) =>
-      memory.subscribe((item) => {
-        callback({
-          timestamp: item.timestamp.toISOString(),
-          conversation_id: item.data.conversationId,
-          query_id: item.data.queryId,
-          message: item.data.message,
-          sequence: item.sequenceNumber,
-        });
-      }),
-    filter: conversationId
-      ? (msg: unknown): boolean =>
-          (msg as {conversation_id: string}).conversation_id === conversationId
-      : undefined,
-    replayItems,
-  });
-}
-
-function handlePaginatedMessages(
-  req: Request,
-  res: Response,
-  memory: MemoryBroker,
-  conversationId: string | undefined,
-  queryId: string | undefined
-): void {
-  try {
-    const params = parsePaginationParams(req.query as Record<string, unknown>);
-
-    const filters = {
-      conversationId: conversationId || undefined,
-      queryId: queryId || undefined,
-    };
-
-    const result = memory.paginate(params, filters);
-
-    const response: PaginatedList<MessageItem> = {
-      items: result.items.map((item) => ({
-        timestamp: item.timestamp.toISOString(),
-        conversation_id: item.data.conversationId,
-        query_id: item.data.queryId,
-        message: item.data.message,
-        sequence: item.sequenceNumber,
-      })),
-      total: result.total,
-      hasMore: result.hasMore,
-      nextCursor: result.nextCursor,
-    };
-
-    res.json(response);
-  } catch (error) {
-    if (error instanceof PaginationError) {
-      sendPaginationError(res, error, req.id);
-      return;
-    }
-    req.log.error({err: error}, 'failed to get messages');
-    sendInternalError(res, req.id);
-  }
-}
+  postMessagesBodySchema,
+  PostMessagesBody,
+  getMessagesQuerySchema,
+  GetMessagesQuery,
+  GetMessagesQueryRaw,
+} from './schemas.js';
+import {handleStreamingMessages, handlePaginatedMessages} from './handlers.js';
 
 export function createMemoryRouter(
   memory: MemoryBroker,
